change for get index in server
Boyang: your program is fine. Can it work? yes it definetly can. But is that logical make sense? I dont really think so. Why? Because to find if the sortedby is legitimate, server could find for me. I'm gonna pass you the index anyway, so you can find it by server and tell me if this is cannot sort or not